Create Directional_Arrow_in_Road_Type_Line_Feature
If a Road feature is bidirectional viz. 2-lane or 4-lane then how can we differentiate with 1-lane in Map. then you must have an Arrow which shows the Direction in Road Feature in it. This Code Chunk explains about how a Coder can able to insert a Arrow with proper direction in different lane road in single shape file unit.